Refactor(soonki/alarm): 폴더 분리
